requirements.txt
CairoSVG~=2.7.1
pillow~=11.2.1
